refactor: export merge utils
